Set up initial page with some questions
Right now this sets up the new FAQ/Options page with most of the
questions for the how to control the selection. Not all of these
questions have answers, most of them weren't moved over, but there are
plans to add detailed answers to all of them.

The questions were aggregated based on common Stack Overflow and IRC
questions, as well as commonly asked things on GitHub.